requirements.txt
requests
python-dotenv
setuptools
motor
beautifulsoup4
playwright
pyrogram
pytz
Flask